+++ include/engine/guidance/route_step.hpp
@@ -49,6 +49,7 @@ struct RouteStep
 {
     unsigned name_id;
     std::string name;
+    std::string pronunciation;
     std::string rotary_name;
     double duration;
     double distance;
@@ -63,6 +64,7 @@ struct RouteStep
 inline RouteStep getInvalidRouteStep()
 {
     return {0,
+            "",
             "",
             "",
             0,